+++ newrelic/data_source_newrelic_alert_policy.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/newrelic/newrelic-client-go/pkg/alerts"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/alerts"
 )
 
 func dataSourceNewRelicAlertPolicy() *schema.Resource {
@@ -47,13 +47,7 @@ func dataSourceNewRelicAlertPolicy() *schema.Resource {
 
 func dataSourceNewRelicAlertPolicyRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	cfg := meta.(*ProviderConfig)
-
-	if !cfg.hasNerdGraphCredentials() {
-		return diag.Errorf("err: NerdGraph support not present, but required for Read")
-	}
-
 	client := cfg.NewClient
-
 	log.Printf("[INFO] Reading New Relic Alert Policies")
 
 	name := d.Get("name").(string)