+++ newrelic/data_source_newrelic_entity.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-	"github.com/newrelic/newrelic-client-go/pkg/entities"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/entities"
 )
 
 func dataSourceNewRelicEntity() *schema.Resource {
@@ -49,7 +49,6 @@ func dataSourceNewRelicEntity() *schema.Resource {
 			},
 			"tag": {
 				Type:        schema.TypeList,
-				MaxItems:    1,
 				Optional:    true,
 				Description: "A tag applied to the entity.",
 				Elem: &schema.Resource{
@@ -117,7 +116,11 @@ func dataSourceNewRelicEntityRead(ctx context.Context, d *schema.ResourceData, m
 	var entity *entities.EntityOutlineInterface
 	for _, e := range entityResults.Results.Entities {
 		// Conditional on case sensitive match
-		if e.GetName() == name || (ignoreCase && strings.EqualFold(e.GetName(), name)) {
+
+		str := e.GetName()
+		str = strings.TrimSpace(str)
+
+		if strings.Compare(str, name) == 0 || (ignoreCase && strings.EqualFold(str, name)) {
 			entity = &e
 			break
 		}