+++ newrelic/provider_unit_test.go
@@ -12,16 +12,3 @@ func TestProvider(t *testing.T) {
 		t.Fatalf("err: %s", err)
 	}
 }
-
-func TestProviderConfig(t *testing.T) {
-	c := ProviderConfig{
-		PersonalAPIKey: "abc123",
-		AccountID:      123,
-	}
-
-	hasNerdGraphCreds := c.hasNerdGraphCredentials()
-
-	if !hasNerdGraphCreds {
-		t.Error("hasNerdGraphCreds should be true")
-	}
-}