+++ newrelic/resource_newrelic_cloud_gcp_integrations.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/newrelic/newrelic-client-go/pkg/cloud"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/cloud"
 )
 
 func resourceNewrelicCloudGcpIntegrations() *schema.Resource {