+++ newrelic/resource_newrelic_notifications_channel.go
@@ -6,13 +6,13 @@ import (
 	"log"
 	"strings"
 
-	"github.com/newrelic/newrelic-client-go/pkg/ai"
-	"github.com/newrelic/newrelic-client-go/pkg/notifications"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/ai"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/notifications"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-	"github.com/newrelic/newrelic-client-go/pkg/errors"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/errors"
 )
 
 func resourceNewRelicNotificationChannel() *schema.Resource {
@@ -86,6 +86,7 @@ func resourceNewRelicNotificationChannelCreate(ctx context.Context, d *schema.Re
 	accountID := selectAccountID(providerConfig, d)
 	updatedContext := updateContextWithAccountID(ctx, accountID)
 	channelInput := expandNotificationChannel(d)
+	channelInput.Properties = append(channelInput.Properties, createMonitoringProperty())
 
 	log.Printf("[INFO] Creating New Relic notification channelResponse %s", channelInput.Name)
 
@@ -134,7 +135,8 @@ func resourceNewRelicNotificationChannelRead(ctx context.Context, d *schema.Reso
 	}
 
 	if len(channelResponse.Entities) == 0 {
-		return diag.FromErr(fmt.Errorf("[ERROR] notification channelResponse.Entities response is empty"))
+		d.SetId("")
+		return nil
 	}
 
 	errors := buildAiNotificationsResponseErrors(channelResponse.Errors)