+++ newrelic/resource_newrelic_one_dashboard_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-	"github.com/newrelic/newrelic-client-go/pkg/common"
-	"github.com/newrelic/newrelic-client-go/pkg/entities"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/common"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/entities"
 )
 
 // TestAccNewRelicOneDashboard_CreateOnePage Ensure that we can create a NR1 Dashboard
@@ -113,7 +113,7 @@ func TestAccNewRelicOneDashboard_PageRename(t *testing.T) {
 			{
 				Config: testAccCheckNewRelicOneDashboardConfig_OnePageFull(rNameUpdated, strconv.Itoa(testAccountID)),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNewRelicOneDashboardExists("newrelic_one_dashboard.bar", 5), // Sleep waiting for entity re-indexing
+					testAccCheckNewRelicOneDashboardExists("newrelic_one_dashboard.bar", 60), // Sleep waiting for entity re-indexing
 				),
 			},
 		},