+++ newrelic/structures_newrelic_alert_condition.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/newrelic/newrelic-client-go/pkg/alerts"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/alerts"
 )
 
 func expandAlertCondition(d *schema.ResourceData) (*alerts.Condition, error) {